backend/requirements_dev.txt
black==22.3.0
django-stubs[compatible-mypy]==1.12.0
isort==5.10.1
mypy==0.961
pylint==2.14.3
pylint-django==2.5.3
types-requests==2.28.0
types-urllib3==1.26.9
